ray casting cache, bug fix for limited angles in super mode.
enhanced mode is strictly superior to standard mode.
super mode is strictly superior to enhanced mode (except for the exclusion of the standard compass point array which I assume is not needed).

super mode appears to only be 20% slower than enhanced. And enhanced is barely measurably slower than standard, and in some cases may be faster than standard.